# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TCF Excellence API",
        "description": "API de la plateforme TCF Excellence - Preparation TCF Canada pour l'Afrique francophone",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/admin/learners": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Liste paginée des apprenants",
                "operationId": "83c864a064e407427539f16bc850e0e7",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "SOLO ou COACHED",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "SOLO",
                                "COACHED"
                            ]
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Filtre par pays",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "description": "Niveau estimé",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "A1",
                                "A2",
                                "B1",
                                "B2",
                                "C1",
                                "C2"
                            ]
                        }
                    },
                    {
                        "name": "inactive",
                        "in": "query",
                        "description": "Inactifs > 7j",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items par page",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste retournée avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/learners/kpis": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "KPIs du tableau de bord admin",
                "operationId": "5aa7c527103bd896eebf111c12bdae7d",
                "responses": {
                    "200": {
                        "description": "KPIs retournés"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/learners/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Profil complet d'un apprenant (drawer)",
                "operationId": "9af377e0a870353eee3dd404a0dcb859",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profil retourné"
                    },
                    "404": {
                        "description": "Apprenant non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/learners/export": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Export CSV des apprenants",
                "operationId": "9d46f65a6c868152cceaaf3de9af676c",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "SOLO",
                                "COACHED"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier CSV téléchargeable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/learners/{id}/message": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Envoyer un message coach à un apprenant",
                "operationId": "66bfc638622d0b5cc17a886c57c6b7b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Bravo pour vos progrès cette semaine !"
                                    },
                                    "is_pinned": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message envoyé"
                    },
                    "404": {
                        "description": "Apprenant non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/stats": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Statistiques globales",
                "operationId": "6098e59d7949d9c2fbdc0ca82cbccf89",
                "responses": {
                    "200": {
                        "description": "Statistiques retournées"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inscription d'un nouvel apprenant",
                "operationId": "ed42f392010c301892e640067141676e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Aissatou Koné"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "aissatou@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+22507000000"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Côte d'Ivoire"
                                    },
                                    "target_exam_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-01"
                                    },
                                    "registration_type": {
                                        "type": "string",
                                        "enum": [
                                            "SOLO",
                                            "COACHED"
                                        ],
                                        "example": "SOLO"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inscription réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Inscription réussie"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123..."
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "aissatou@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123..."
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Identifiants incorrects"
                    },
                    "403": {
                        "description": "Compte désactivé"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnexion",
                "operationId": "0d29c740dc06d77602af8095bf4654cc",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Profil de l'utilisateur connecté",
                "operationId": "a9c495788506b114956e3f509e5d90a2",
                "responses": {
                    "200": {
                        "description": "Profil retourné avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Admin",
            "description": "Admin"
        },
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}